Move utilities
Prompt: "Move list_files and synchronize_files to the utils.py file"
